requirements.txt
python-dotenv
langchain
pinecone-client
langchain-pinecone
langchain-openai
pypdf
streamlit